Hiking Habits Application Project
This project serves to display use of software development techniques in order to develop an application that serves as a social media for people who want to view submitted locations for biking, hiking, walking, etc. In this application, users submit locations they've been to with relevant info, and then they can also search for other submissions from other people. This project uses MySQL to connect to a local database containing relevant info.